requirements.txt
# =============================================================================
# REAP V11 - Python Dependencies
# =============================================================================

# PyTorch (install with CUDA separately)
torch>=2.4.0
torchvision>=0.19.0

# Transformers
transformers>=4.45.0
accelerate>=0.34.0
safetensors>=0.4.5
huggingface-hub>=0.25.0
tokenizers>=0.19.0
sentencepiece>=0.2.0

# Flash Attention
flash-attn>=2.6.0

# Scientific
numpy>=1.26.0
scipy>=1.13.0

# Image
Pillow>=10.4.0

# Qwen-VL
qwen-vl-utils>=0.0.8

# Progress
tqdm>=4.66.0

# Utils
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
vllm>=0.6.0
